Refactored existing skupper link status to promote reuse
